+++ src/chess/bitboard.h
@@ -32,11 +32,6 @@
 #include <string>
 #include <vector>
 
-#if not defined(NO_PEXT)
-// Include header for pext instruction.
-#include <immintrin.h>
-#endif
-
 #include "utils/bititer.h"
 
 namespace lczero {
@@ -236,86 +231,6 @@ class BitBoard {
 
  private:
   std::uint64_t board_ = 0;
-};
-
-// Holds magic bitboard routines and performs initialization.
-// We use so-called "fancy" magic bitboards.
-class MagicBitBoards {
- public:
-  // Constructor is called once in bitboard.cc to initialize structures, should
-  // not be called again!
-  MagicBitBoards();
-
-  // Structure holding all relevant magic parameters per square.
-  struct MagicParams {
-    // Relevant occupancy mask.
-    uint64_t mask_;
-    // Pointer to lookup table.
-    BitBoard* attacks_table_;
-#if defined(NO_PEXT)
-    // Magic number.
-    uint64_t magic_number_;
-    // Number of bits to shift.
-    uint8_t shift_bits_;
-#endif
-  };
-
-  // Returns the rook attacks bitboard for the given rook board square and the
-  // given occupied piece bitboard.
-  static BitBoard GetRookAttacks(const BoardSquare rook_square,
-                                 const BitBoard pieces) {
-    // Calculate magic index.
-    const uint8_t square = rook_square.as_int();
-
-#if defined(NO_PEXT)
-    uint64_t index = pieces.as_int() & rook_magic_params_[square].mask_;
-    index *= rook_magic_params_[square].magic_number_;
-    index >>= rook_magic_params_[square].shift_bits_;
-#else
-    uint64_t index =
-        _pext_u64(pieces.as_int(), rook_magic_params_[square].mask_);
-#endif
-
-    // Return attacks bitboard.
-    return rook_magic_params_[square].attacks_table_[index];
-  }
-
-  // Returns the bishop attacks bitboard for the given bishop board square and
-  // the given occupied piece bitboard.
-  static BitBoard GetBishopAttacks(const BoardSquare bishop_square,
-                                   const BitBoard pieces) {
-    // Calculate magic index.
-    const uint8_t square = bishop_square.as_int();
-
-#if defined(NO_PEXT)
-    uint64_t index = pieces.as_int() & bishop_magic_params_[square].mask_;
-    index *= bishop_magic_params_[square].magic_number_;
-    index >>= bishop_magic_params_[square].shift_bits_;
-#else
-    uint64_t index =
-        _pext_u64(pieces.as_int(), bishop_magic_params_[square].mask_);
-#endif
-
-    // Return attacks bitboard.
-    return bishop_magic_params_[square].attacks_table_[index];
-  }
-
- private:
-  // Builds rook or bishop attacks table.
-  void BuildAttacksTable(MagicParams* magic_params, BitBoard* attacks_table,
-                         const std::pair<int, int>* directions);
-
-  // Magic numbers for each board square.
-  static const BitBoard kRookMagicNumbers[64];
-  static const BitBoard kBishopMagicNumbers[64];
-
-  // Magic parameters for each board square.
-  static MagicParams rook_magic_params_[64];
-  static MagicParams bishop_magic_params_[64];
-
-  // Attacks bitboards lookup tables.
-  static BitBoard rook_attacks_table_[102400];
-  static BitBoard bishop_attacks_table_[5248];
 };
 
 class Move {
